chore(alerts): refactor code and cleanup
