+++ packages/blockchain-wallet-v4-frontend/src/components/Alerts/index.tsx
@@ -1,25 +1,68 @@
 import React from 'react'
-import { connect, ConnectedProps } from 'react-redux'
-import { bindActionCreators } from 'redux'
+import { useDispatch, useSelector } from 'react-redux'
+import styled, { keyframes } from 'styled-components'
 
-import { actions, selectors } from 'data'
+import { Toast } from 'blockchain-info-components'
+import { selectAlerts } from 'data/alerts/selectors'
+import { actions as alertsActions } from 'data/alerts/slice'
+import { media } from 'services/styles'
 
-import Alerts from './template'
+import getAlertContent from './messages'
 
-const AlertsContainer = ({ alertActions, alerts }: Props) => {
-  return <Alerts alerts={alerts || []} handleClose={(id) => alertActions.dismissAlert(id)} />
+const Wrapper = styled.div`
+  position: absolute;
+  display: flex;
+  flex-direction: column-reverse;
+  justify-content: flex-start;
+  align-items: flex-start;
+  z-index: 1050;
+
+  ${media.tablet`
+    bottom: 10px;
+    right: 10px;
+    width: 95%;
+  `}
+  ${media.atLeastTablet`
+    top: 65px;
+    right: 22px;
+    width: auto;
+  `}
+`
+const easeIn = keyframes`
+  0% {opacity: 0;}
+  100% {opacity: 1;}
+`
+const AnimatedToast = styled.div`
+  animation-name: ${easeIn};
+  animation-duration: 0.5s;
+`
+
+const DISMISS_ALERT_TIME = 5000
+
+const Alerts = () => {
+  const alerts = useSelector(selectAlerts)
+  const dispatch = useDispatch()
+
+  return (
+    <Wrapper>
+      {alerts.map((alert) => {
+        const { coin, data, id, message, nature, persist } = alert
+        return (
+          <AnimatedToast key={id}>
+            <Toast
+              coin={coin}
+              nature={nature}
+              onClose={() => dispatch(alertsActions.dismissAlert(id))}
+              persist={persist}
+              timeout={DISMISS_ALERT_TIME}
+            >
+              {getAlertContent(message, data)}
+            </Toast>
+          </AnimatedToast>
+        )
+      })}
+    </Wrapper>
+  )
 }
 
-const mapStateToProps = (state) => ({
-  alerts: selectors.alerts.selectAlerts(state)
-})
-
-const mapDispatchToProps = (dispatch) => ({
-  alertActions: bindActionCreators(actions.alerts, dispatch)
-})
-
-const connector = connect(mapStateToProps, mapDispatchToProps)
-
-type Props = ConnectedProps<typeof connector>
-
-export default connector(AlertsContainer)
+export default Alerts
