--- packages/blockchain-wallet-v4-frontend/src/data/alerts/sagaRegister.ts
@@ -1,8 +1,0 @@
-import { takeEvery } from 'redux-saga/effects'
-
-import * as sagas from './sagas'
-import { showAlerts } from './slice'
-
-export default function* alertsSaga() {
-  yield takeEvery(showAlerts.type, sagas.handleTimer)
-}
