+++ packages/blockchain-wallet-v4-frontend/src/data/alerts/slice.ts
@@ -1,12 +1,11 @@
-import { createAction, createSlice } from '@reduxjs/toolkit'
-import { prepend } from 'ramda'
+import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
 import { CoinType } from '@core/types'
 
-import { AlertNatureType, AlertsState } from './types'
+import { AlertNatureType, AlertsState, AlertType } from './types'
 
 const initialState: AlertsState = []
-const generateId = () => Math.random().toString(36).substr(2, 10)
+const generateId = () => Math.random().toString(36).substring(2, 10)
 
 const alertsSlice = createSlice({
   initialState,
@@ -18,38 +17,34 @@ const alertsSlice = createSlice({
     dismissAlert: (state, action) => {
       return state.filter((alert) => alert.id !== action.payload)
     },
-    showAlerts: (state, action) => {
-      return prepend({ ...action.payload }, state)
+    showAlerts: {
+      prepare: (
+        nature: AlertNatureType,
+        message: string,
+        data?: any,
+        persist?: boolean,
+        coin?: CoinType
+      ) => {
+        return { payload: { coin, data, id: generateId(), message, nature, persist } }
+      },
+      reducer: (state, action: PayloadAction<AlertType>) => {
+        return state.concat(action.payload)
+      }
     }
   }
 })
 
-const display = createAction(
-  'alerts/showAlerts',
-  function prepare(
-    nature: AlertNatureType,
-    message: string,
-    data: any,
-    persist?: boolean,
-    coin?: CoinType
-  ) {
-    return {
-      payload: { coin, data, id: generateId(), message, nature, persist }
-    }
-  }
-)
-
 const displayWarning = (message: string, data?: any, persist?: boolean) =>
-  display('warn', message, data, persist)
+  alertsSlice.actions.showAlerts('warn', message, data, persist)
 
 const displayInfo = (message: string, data?: any, persist?: boolean) =>
-  display('info', message, data, persist)
+  alertsSlice.actions.showAlerts('info', message, data, persist)
 
 const displaySuccess = (message: string, data?: any, persist?: boolean) =>
-  display('success', message, data, persist)
+  alertsSlice.actions.showAlerts('success', message, data, persist)
 
 const displayError = (message: string, data?: any, persist?: boolean, coin?: CoinType) =>
-  display('error', message, data, persist, coin)
+  alertsSlice.actions.showAlerts('error', message, data, persist, coin)
 
 export const actions = {
   ...alertsSlice.actions,
@@ -58,8 +53,5 @@ export const actions = {
   displaySuccess,
   displayWarning
 }
-export const { clearAlerts, dismissAlert, showAlerts } = actions
 
 export const alertsReducer = alertsSlice.reducer
-
-export { displayError, displayInfo, displaySuccess, displayWarning }
